// FE/src/app/projects/w-property/interfaces/request.interface.ts
export interface IWPropertyRequest {
  name?: string;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  userCountry: string;
  userCity: string;
  userInfo: string;
  userUnits: string;
  unitStatus: string;
  rentType: string;
  promoCode: string;
  userRelation: string;
  isInvestBefore: string;
  callAgreement: string;
  unitCountry: string;
  unitCity: string;
  unitDistrict: string;
  unitInfo: string;
}